match apis

+++ klondike/bigquery/bigquery.py
@@ -9,6 +9,7 @@ from google.cloud import bigquery
 from google.cloud.bigquery import LoadJobConfig
 
 from klondike import logger
+from klondike.utilities.utilities import validate_if_exists_behavior
 
 ##########
 
@@ -26,12 +27,18 @@ class BigQueryConnector:
     Establish and authenticate a connection to a BigQuery warehouse
 
     Args:
-        app_creds: Google service account, either as a relative path or a dictionary instance
-        project: Name of Google Project
-        location: Location of Google Project
-        timeout: Temporal threshold to kill a stalled job, defaults to 60s
-        client_options: API scopes
-        google_environment_variable: Provided for flexibility, defaults to `GOOGLE_APPLICATION_CREDENTIALS`
+        app_creds: `str`
+            Google service account, either as a relative path or a dictionary instance
+        project: `str`
+            Name of Google Project
+        location: `str`
+            Location of Google Project
+        timeout: `int`
+            Temporal threshold to kill a stalled job, defaults to 60s
+        client_options: `list`
+            API scopes
+        google_environment_variable: `str`
+            Provided for flexibility, defaults to `GOOGLE_APPLICATION_CREDENTIALS`
     """
 
     def __init__(
@@ -46,11 +53,11 @@ class BigQueryConnector:
         self.app_creds = app_creds
         self.project = project
         self.location = location
-        self.timeout = timeout
         self.client_options = client_options
 
-        self._client = None
         self.dialect = "bigquery"
+        self.__client = None
+        self.__timeout = timeout
 
         if not self.app_creds:
             if not os.environ.get(google_environment_variable):
@@ -63,6 +70,30 @@ class BigQueryConnector:
             self.__setup_google_app_creds(
                 app_creds=self.app_creds, env_variable=google_environment_variable
             )
+
+    @property
+    def client(self):
+        """
+        Instantiate BigQuery client and assign it
+        as class property
+        """
+
+        if not self.__client:
+            self.__client = bigquery.Client(
+                project=self.project,
+                location=self.location,
+                client_options=self.client_options,
+            )
+
+        return self.__client
+
+    @property
+    def timeout(self):
+        return self.__timeout
+
+    @timeout.setter
+    def timeout(self, timeout):
+        self.__timeout = timeout
 
     def __setup_google_app_creds(self, app_creds: Union[str, dict], env_variable: str):
         "Sets runtime environment variable for Google SDK"
@@ -89,6 +120,20 @@ class BigQueryConnector:
     ):
         "Defines `LoadConfigJob` when writing to BigQuery"
 
+        def set_write_disposition(if_exists: str):
+            DISPOSITION_MAP = {
+                "fail": bigquery.WriteDisposition.WRITE_EMPTY,
+                "append": bigquery.WriteDisposition.WRITE_APPEND,
+                "truncate": bigquery.WriteDisposition.WRITE_TRUNCATE,
+            }
+
+            return DISPOSITION_MAP[if_exists]
+
+        def set_table_schema(table_schema: list):
+            return [bigquery.SchemaField(**x) for x in table_schema]
+
+        ###
+
         if not base_job_config:
             logger.debug("No job config provided, starting fresh")
             base_job_config = LoadJobConfig()
@@ -98,17 +143,16 @@ class BigQueryConnector:
 
         # Create table schema mapping if provided
         if table_schema:
-            base_job_config.schema = self.__set_table_schema(table_schema=table_schema)
+            base_job_config.schema = set_table_schema(table_schema=table_schema)
         else:
             base_job_config.schema = None
 
         base_job_config.max_bad_records = max_bad_records
+        base_job_config.write_disposition = set_write_disposition(if_exists=if_exists)
 
-        base_job_config.write_disposition = self.__set_write_disposition(
-            if_exists=if_exists
-        )
+        ###
 
-        # List of LoadJobConfig attributes
+        # List of available LoadJobConfig attributes
         _attributes = [x for x in dict(vars(LoadJobConfig)).keys()]
 
         # Attributes that will not be overwritten
@@ -127,44 +171,14 @@ class BigQueryConnector:
 
         return base_job_config
 
-    def __set_table_schema(self, table_schema: list):
-        "TODO - Write about me"
-
-        return [bigquery.SchemaField(**x) for x in table_schema]
-
-    def __set_write_disposition(self, if_exists: str):
-        "TODO - Write about me"
-
-        DISPOSITION_MAP = {
-            "fail": bigquery.WriteDisposition.WRITE_EMPTY,
-            "append": bigquery.WriteDisposition.WRITE_APPEND,
-            "truncate": bigquery.WriteDisposition.WRITE_TRUNCATE,
-        }
-
-        return DISPOSITION_MAP[if_exists]
-
-    @property
-    def client(self):
-        """
-        Instantiate BigQuery client
-        """
-
-        if not self._client:
-            self._client = bigquery.Client(
-                project=self.project,
-                location=self.location,
-                client_options=self.client_options,
-            )
-
-        return self._client
-
-    def read_dataframe_from_bigquery(self, sql: str) -> pl.DataFrame:
+    def read_dataframe(self, sql: str) -> pl.DataFrame:
         """
         Executes a SQL query and returns a Polars DataFrame.
         TODO - Make this more flexible and incorporate query params
 
         Args:
-            sql: String representation of SQL query
+            sql: `str`
+                String representation of SQL query
 
         Returns:
             Polars DataFrame object
@@ -185,9 +199,10 @@ class BigQueryConnector:
             return
 
         logger.info(f"Successfully read {len(df)} rows from BigQuery")
+
         return df
 
-    def write_dataframe_to_bigquery(
+    def write_dataframe(
         self,
         df: pl.DataFrame,
         table_name: str,
@@ -210,6 +225,9 @@ class BigQueryConnector:
             load_kwargs: See here for list of accepted values
                 https://cloud.google.com/python/docs/reference/bigquery/latest/google.cloud.bigquery.job.LoadJobConfig
         """
+
+        if not validate_if_exists_behavior(user_input=if_exists):
+            raise ValueError(f"{if_exists} is an invalid input")
 
         if if_exists == "drop":
             self.client.delete_table(table=table_name)
